+++ packages/core/src/model/node/BaseNodeModel.ts
@@ -19,6 +19,7 @@ import {
   IsAllowMove,
   Point,
   AnchorInfo,
+  DomAttributes,
 } from '../../type';
 import GraphModel from '../GraphModel';
 import { IBaseModel } from '../BaseModel';
@@ -26,6 +27,7 @@ import { formatData } from '../../util/compatible';
 import { getClosestAnchor, pickNodeConfig } from '../../util/node';
 import { getZIndex } from '../../util/zIndex';
 import { BaseEdgeModel } from '../edge';
+import { Matrix, TranslateMatrix } from '../../util';
 
 export type ConnectRule = {
   message: string;
@@ -89,6 +91,7 @@ export default class BaseNodeModel implements IBaseNodeModel {
   @observable isHitable = true; // 细粒度控制节点是否对用户操作进行反应
   @observable draggable = true;
   @observable visible = true;
+  @observable enableRotate = true;
   virtual = false;
   // 其它属性
   graphModel: GraphModel;
@@ -96,6 +99,17 @@ export default class BaseNodeModel implements IBaseNodeModel {
   @observable state = 1;
   @observable autoToFront = true; // 节点选中时是否自动置顶，默认为true.
   @observable style: ShapeStyleAttribute = { }; // 每个节点自己的样式，动态修改
+  @observable transform: string; // 节点的transform属性
+  @observable private _rotate = 0;
+  set rotate(value: number) {
+    this._rotate = value;
+    const { x = 0, y = 0 } = this;
+    this.transform = new TranslateMatrix(-x, -y).rotate(value).translate(x, y).toString();
+  }
+  get rotate() {
+    return this._rotate;
+  }
+
   readonly BaseType = ElementType.NODE;
   modelType = ModelType.NODE;
   additionStateData: AdditionData;
@@ -218,6 +232,9 @@ export default class BaseNodeModel implements IBaseNodeModel {
       y: this.y,
       properties,
     };
+    if (this.rotate) {
+      data.rotate = this.rotate;
+    }
     if (this.graphModel.overlapMode === OverlapMode.INCREASE) {
       data.zIndex = this.zIndex;
     }
@@ -246,6 +263,16 @@ export default class BaseNodeModel implements IBaseNodeModel {
   }
   /**
    * @overridable 支持重写
+   * 获取当前节点最外层g标签Attributes, 例如className
+   * @returns 自定义节点样式
+   */
+  getOuterGAttributes(): DomAttributes {
+    return {
+      className: '',
+    };
+  }
+  /**
+   * @overridable 支持重写
    * 获取当前节点样式
    * @returns 自定义节点样式
    */
@@ -263,6 +290,14 @@ export default class BaseNodeModel implements IBaseNodeModel {
     // 透传 nodeText
     const { nodeText } = this.graphModel.theme;
     return cloneDeep(nodeText);
+  }
+  /**
+   * @overridable 支持重写
+   * 获取当前节点旋转控制点的样式
+   */
+  getRotateControlStyle() {
+    const { rotateControl } = this.graphModel.theme;
+    return cloneDeep(rotateControl);
   }
   /**
    * @overridable 支持重写
@@ -454,7 +489,18 @@ export default class BaseNodeModel implements IBaseNodeModel {
   }
 
   get anchors(): PointAnchor[] {
-    return this.getAnchorsByOffset();
+    const anchors = this.getAnchorsByOffset();
+    const { x, y, rotate } = this;
+    anchors.forEach((anchor) => {
+      const { x: anchorX, y: anchorY } = anchor;
+      const [e, f] = new Matrix([anchorX, anchorY, 1])
+        .translate(-x, -y)
+        .rotate(rotate)
+        .translate(x, y)[0];
+      anchor.x = e;
+      anchor.y = f;
+    });
+    return anchors;
   }
 
   getAnchorInfo(anchorId: string) {
@@ -589,6 +635,11 @@ export default class BaseNodeModel implements IBaseNodeModel {
   @action
   setIsShowAnchor(flag = true): void {
     this.isShowAnchor = flag;
+  }
+
+  @action
+  setEnableRotate(flag = true): void {
+    this.enableRotate = flag;
   }
 
   @action