+++ packages/core/src/util/index.ts
@@ -1,2 +1,7 @@
 export * from './uuid';
 export * from './drag';
+export * from './edge';
+export * from './geometry';
+export * from './sampling';
+export * from './vector';
+export * from './matrix';