+++ packages/engine/__test__/02_recorder.test.js
@@ -1,8 +1,60 @@
 import Engine, { Recorder } from '../src/index';
 
 describe('@logicflow/engine Recorder', () => {
+  test('When init Engine with debug mode, getExecutionRecord will get null', async () => {
+    const engine = new Engine({
+      debug: false
+    });
+    const flowData = {
+      /**
+       * node1 |--> node2
+       */
+      graphData: {
+        nodes: [
+          {
+            id: 'node1',
+            type: 'StartNode',
+            properties: {}
+          },
+          {
+            id: 'node2',
+            type: 'TaskNode',
+            properties: {}
+          }
+        ],
+        edges: [
+          {
+            id: 'edge1',
+            sourceNodeId: 'node1',
+            targetNodeId: 'node2',
+          }
+        ]
+      },
+      global: {},
+    }
+    engine.load(flowData);
+    const result = await engine.execute();
+    const executionId = result.executionId;
+    /**
+     * [
+     *   {
+     *     actionId: '', // 某一个节点在某一次执行时生成的Id
+     *     nodeId: '', // 流程图节点Id
+     *     executionId: '', // 某一次执行的Id
+     *     nodeType: '',
+     *     timestamp: '',
+     *     properties: {},
+     *   }
+     * ]
+     */
+    // TODO: 给个例子自定义执行记录
+    const execution = await engine.getExecutionRecord(executionId);
+    expect(execution).toBe(null)
+  });
   test('Using the getExecutionRecord API, receive the complete execution record of the process.', async () => {
-    const engine = new Engine();
+    const engine = new Engine({
+      debug: true,
+    });
     const flowData = {
       /**
        * node1 |--> node2
@@ -60,7 +112,9 @@ describe('@logicflow/engine Recorder', () => {
     expect(executionIds.length).toBe(1);
   });
   test('The execution record cannot be obtained when the number of executions exceeds the maximum number of executions.', async () => {
-    const engine = new Engine();
+    const engine = new Engine({
+      debug: true,
+    });
     const flowData = {
       /**
        * node1 |--> node2
@@ -124,11 +178,15 @@ describe('@logicflow/engine Recorder', () => {
       },
       global: {},
     }
-    const engine = new Engine();
+    const engine = new Engine({
+      debug: true,
+    });
     engine.load(flowData);
     await engine.execute();
     await engine.execute();
-    const engine1 = new Engine();
+    const engine1 = new Engine({
+      debug: true,
+    });
     engine1.load(JSON.parse(JSON.stringify(flowData)));
     await engine1.execute();