+++ packages/engine/__test__/03_condition.test.js
@@ -2,7 +2,9 @@ import Engine from '../src/index';
 
 describe('@logicflow/engine condition', () => {
   test('The process will not continue its execution if the condition expression evaluates to false.', async () => {
-    const engine = new Engine();
+    const engine = new Engine({
+      debug: true,
+    });
     const flowData = {
       /**
        * node1 |--> node2
@@ -56,5 +58,7 @@ describe('@logicflow/engine condition', () => {
     expect(execution.length).toBe(2);
     expect(execution[1].nodeId).toBe('node3');
     expect(execution[1].nodeType).toBe('TaskNode');
+    expect(execution[0].outgoing[0].result).toBe(false);
+    expect(execution[0].outgoing[1].result).toBe(true);
   });
 });