+++ packages/engine/__test__/05_customNode.test.js
@@ -6,6 +6,12 @@ describe('@logicflow/engine Customize Node', () => {
       this.globalData['dataSource'] = {
         time: this.context.getTime(),
       }
+      return {
+        status: 'success',
+        detail: {
+          customData: '2'
+        }
+      }
     }
   }
   class Mod2Node extends TaskNode {
@@ -27,7 +33,8 @@ describe('@logicflow/engine Customize Node', () => {
       getTime() {
         return new Date().getTime();
       }
-    }
+    },
+    debug: true,
   });
   engine.register({
     type: 'DataNode',
@@ -99,8 +106,17 @@ describe('@logicflow/engine Customize Node', () => {
   }
   engine.load(flowData);
   test('When the process is completed, the output field in the properties attribute of the last node is odd or even.', async () => {
-    const result = await engine.execute();
+    const result = await engine.execute({
+      debug: true,
+    });
     const execution = await engine.getExecutionRecord(result.executionId);
     expect(['odd', 'even'].indexOf(execution[execution.length - 1].properties.output) !== -1).toBe(true);
   });
+  test('Execution records will contain return detail', async () => {
+    const result = await engine.execute({
+      debug: true,
+    });
+    const execution = await engine.getExecutionRecord(result.executionId);
+    expect(execution[1].detail.customData).toBe('2')
+  })
 });