+++ packages/engine/src/nodes/BaseNode.ts
@@ -1,10 +1,11 @@
 import { ActionStatus } from '../constant/constant';
 import { getExpressionResult } from '../expression';
 import type {
+  NextActionParam,
   ActionResult,
-  NodeExecResult,
   ExecResumeParams,
   ExecParams,
+  OutgoingConfig,
 } from '../types.d';
 
 export interface BaseNodeInterface {
@@ -13,7 +14,7 @@ export interface BaseNodeInterface {
   nodeId: string;
   type: string;
   readonly baseType: string;
-  execute(actionParam): Promise<NodeExecResult>;
+  execute(actionParam): Promise<NextActionParam>;
 }
 
 export type NodeConstructor = {
@@ -26,7 +27,7 @@ export type NodeConstructor = {
     executionId: string;
     actionId: string;
     nodeId: string;
-  }): Promise<ActionResult>;
+  }): Promise<NextActionParam>;
   onResume(params: {
     executionId: string;
     actionId: string;
@@ -41,27 +42,12 @@ export type IncomingConfig = {
   source: string;
 };
 
-export type OutgoingConfig = {
-  id: string;
-  target: string;
-  properties?: Record<string, any>;
-};
-
 export type NodeConfig = {
   id: string;
   type: string;
   properties?: Record<string, any>;
   incoming: IncomingConfig[];
   outgoing: OutgoingConfig[];
-};
-
-export type NextActionParam = {
-  executionId: string;
-  nodeId: string;
-  actionId: string;
-  nodeType: string;
-  outgoing: OutgoingConfig[];
-  properties?: Record<string, any>;
 };
 
 export default class BaseNode implements BaseNodeInterface {
@@ -103,15 +89,19 @@ export default class BaseNode implements BaseNodeInterface {
   /**
    * 节点的每一次执行都会生成一个唯一的actionId
    */
-  public async execute(params: ExecParams): Promise<NodeExecResult> {
+  public async execute(params: ExecParams): Promise<NextActionParam> {
     const r = await this.action({
       executionId: params.executionId,
       actionId: params.actionId,
       nodeId: this.nodeId,
     });
-    if (!r || r.status === ActionStatus.SUCCESS) {
+    const status = r ? r.status : 'success';
+    if (status === ActionStatus.SUCCESS) {
       const outgoing = await this.getOutgoing();
+      const detail = r ? r.detail : {};
       params.next({
+        status: ActionStatus.SUCCESS,
+        detail,
         executionId: params.executionId,
         actionId: params.actionId,
         nodeId: this.nodeId,
@@ -121,13 +111,14 @@ export default class BaseNode implements BaseNodeInterface {
       });
     }
     return {
-      status: r && r.status,
+      status,
       detail: r && r.detail,
       executionId: params.executionId,
       actionId: params.actionId,
       nodeId: this.nodeId,
       nodeType: this.type,
       properties: this.properties,
+      outgoing: [],
     };
   }
   /**
@@ -149,11 +140,12 @@ export default class BaseNode implements BaseNodeInterface {
       nodeType: this.type,
       properties: this.properties,
       outgoing,
+      status: ActionStatus.SUCCESS,
     });
     return undefined;
   }
   private async getOutgoing(): Promise<OutgoingConfig[]> {
-    const outgoing = [];
+    const outgoing: OutgoingConfig[] = [];
     const expressions = [];
     for (const item of this.outgoing) {
       const { properties } = item;
@@ -161,9 +153,9 @@ export default class BaseNode implements BaseNodeInterface {
     }
     const result = await Promise.all(expressions);
     result.forEach((item, index) => {
-      if (item) {
-        outgoing.push(this.outgoing[index]);
-      }
+      const out = this.outgoing[index];
+      out.result = item;
+      outgoing.push(out);
     });
     return outgoing;
   }
@@ -172,10 +164,12 @@ export default class BaseNode implements BaseNodeInterface {
     const { conditionExpression } = properties;
     if (!conditionExpression) return true;
     try {
-      const result = await getExpressionResult(`result${this.nodeId} = (${conditionExpression})`, {
+      // bug：uuid 创建的 NodeId 为 xxxx-xxxx-xxxx-zzzz 格式，eval 执行时会将 - 识别为数学减号，导致执行报错
+      // 解决方案： 赋值变量直接命名为 isPassResult, 因为每次执行 getExpressionResult 时，都会重新射程一个 context
+      const result = await getExpressionResult(`isPassResult = (${conditionExpression})`, {
         ...this.globalData,
       });
-      return result[`result${this.nodeId}`];
+      return result.isPassResult;
     } catch (e) {
       return false;
     }
@@ -186,13 +180,16 @@ export default class BaseNode implements BaseNodeInterface {
    * @param params.executionId 流程执行记录ID
    * @param params.actionId 此节点执行记录ID
    * @param params.nodeId 节点ID
+   * @returns 返回下一步的执行参数
+   * 当不返回时，表示此节点执行成功，流程会继续执行下一步。
+   * 当返回时，返回格式
    */
   public async action(params: {
     executionId: string;
     actionId: string;
     nodeId: string;
   }): Promise<ActionResult> {
-    return undefined;
+    return null;
   }
   /**
    * 节点的重新恢复执行逻辑